+++ sams/output/Collectd.py
@@ -50,7 +50,7 @@ logger = logging.getLogger(__name__)
 
 
 class Output(sams.base.Output):
-    """ File output Class """
+    """File output Class"""
 
     def __init__(self, id, config):
         super(Output, self).__init__(id, config)